+++ tests/test_write_path_eager_loading.py
@@ -110,10 +110,10 @@ def _build_nested_app(app):
         schema = DocRead
         schema_create = DocCreate
 
-        async def make_new_object(self, schema_obj) -> Any:
-            obj: Any = await super().make_new_object(schema_obj)
+        async def create(self, schema_obj) -> Any:
+            obj: Any = await self.make_new_object(schema_obj)
             obj.owner_id = owner_id["id"]
-            return obj
+            return await self.save_object(obj)
 
     create_tables()
 
@@ -177,10 +177,10 @@ def test_sync_create_serializes_nested_relationship(sync_client):
         schema = DocRead
         schema_create = DocCreate
 
-        def make_new_object(self, schema_obj) -> Any:
-            obj: Any = super().make_new_object(schema_obj)
+        def create(self, schema_obj) -> Any:
+            obj: Any = self.make_new_object(schema_obj)
             obj.owner_id = owner_id["id"]
-            return obj
+            return self.save_object(obj)
 
     _create_sync_tables()
     with _fr_globals.make_session() as session:
@@ -277,11 +277,11 @@ def test_async_create_serializes_assigned_collection(client):
         schema = DocRead
         schema_create = DocCreate
 
-        async def make_new_object(self, schema_obj) -> Any:
-            obj: Any = await super().make_new_object(schema_obj)
+        async def create(self, schema_obj) -> Any:
+            obj: Any = await self.make_new_object(schema_obj)
             tags = (await self.session.scalars(select(Tag))).all()
             obj.tags = list(tags)
-            return obj
+            return await self.save_object(obj)
 
     create_tables()
 
@@ -428,10 +428,10 @@ def test_indirection_over_a_named_relationship_rides_along(client):
         schema = DocRead
         schema_create = DocCreate
 
-        async def make_new_object(self, schema_obj) -> Any:
-            obj: Any = await super().make_new_object(schema_obj)
+        async def create(self, schema_obj) -> Any:
+            obj: Any = await self.make_new_object(schema_obj)
             obj.owner_id = owner_id["id"]
-            return obj
+            return await self.save_object(obj)
 
     create_tables()
 
@@ -488,10 +488,10 @@ def test_indirection_over_an_unnamed_relationship_fails_alike_on_read_and_write(
         schema = DocRead
         schema_create = DocCreate
 
-        async def make_new_object(self, schema_obj) -> Any:
-            obj: Any = await super().make_new_object(schema_obj)
+        async def create(self, schema_obj) -> Any:
+            obj: Any = await self.make_new_object(schema_obj)
             obj.owner_id = 1
-            return obj
+            return await self.save_object(obj)
 
     create_tables()
 
@@ -657,12 +657,12 @@ def test_divergent_schemas_same_instance_reload(client):
         schema = DocRead
         schema_create = DocCreate
 
-        async def make_new_object(self, schema_obj) -> Any:
-            obj: Any = await super().make_new_object(schema_obj)
+        async def create(self, schema_obj) -> Any:
+            obj: Any = await self.make_new_object(schema_obj)
             # author and reviewer are the *same* person row
             obj.author_id = 1
             obj.reviewer_id = 1
-            return obj
+            return await self.save_object(obj)
 
     create_tables()
 
@@ -722,10 +722,10 @@ def test_sync_create_serializes_assigned_collection(sync_client):
         schema = DocRead
         schema_create = DocCreate
 
-        def make_new_object(self, schema_obj) -> Any:
-            obj: Any = super().make_new_object(schema_obj)
+        def create(self, schema_obj) -> Any:
+            obj: Any = self.make_new_object(schema_obj)
             obj.tags = list(self.session.scalars(select(Tag)).all())
-            return obj
+            return self.save_object(obj)
 
         def to_response(self, obj_or_list, *args, **kwargs) -> Any:
             # snapshot load state after save_object, before serialization reads it
@@ -786,10 +786,10 @@ def test_write_reload_unique_guards_a_joinedload_collection(client):
             # .all() would raise InvalidRequestError
             return [joinedload(JoinedDoc.tags)]
 
-        async def make_new_object(self, schema_obj) -> Any:
-            obj: Any = await super().make_new_object(schema_obj)
+        async def create(self, schema_obj) -> Any:
+            obj: Any = await self.make_new_object(schema_obj)
             obj.tags = list((await self.session.scalars(select(Tag))).all())
-            return obj
+            return await self.save_object(obj)
 
     create_tables()